Rename Cell to CoreCell (cell.py -> core_cell.py) Rename NewCell to Cell (new_cell.py -> cell.py)
Remove comment and unused codes
