+++ nvflare/fuel/f3/cellnet/cbs.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from nvflare.fuel.f3.cellnet.cell import CellAgent
+from nvflare.fuel.f3.cellnet.core_cell import CellAgent
 from nvflare.fuel.f3.message import Message
