+++ nvflare/private/fed/client/fed_client.py
@@ -22,13 +22,11 @@ from nvflare.apis.filter import Filter
 from nvflare.apis.fl_component import FLComponent
 from nvflare.apis.fl_constant import FLContextKey
 from nvflare.apis.fl_context import FLContext
-from nvflare.fuel.f3.cellnet.new_cell import NewCell as Cell
+from nvflare.fuel.f3.cellnet.cell import Cell
 from nvflare.private.defs import SpecialTaskName
 from nvflare.private.event import fire_event
 
 from .fed_client_base import FederatedClientBase
-
-# from nvflare.fuel.f3.cellnet.cell import Cell
 
 
 class FederatedClient(FederatedClientBase):
