+++ nvflare/private/fed/server/server_engine.py
@@ -46,7 +46,7 @@ from nvflare.apis.job_def import Job
 from nvflare.apis.shareable import Shareable
 from nvflare.apis.utils.fl_context_utils import get_serializable_data
 from nvflare.apis.workspace import Workspace
-from nvflare.fuel.f3.cellnet.cell import FQCN, Cell
+from nvflare.fuel.f3.cellnet.core_cell import FQCN, CoreCell
 from nvflare.fuel.f3.cellnet.defs import MessageHeaderKey
 from nvflare.fuel.f3.cellnet.defs import ReturnCode as CellMsgReturnCode
 from nvflare.fuel.utils import fobs
@@ -244,7 +244,7 @@ class ServerEngine(ServerEngineInternalSpec):
         job_id,
         job_clients,
         snapshot,
-        cell: Cell,
+        cell: CoreCell,
         server_state: ServerState,
     ):
         new_env = os.environ.copy()
